Square bracket bug in MWC count was fixed.
